SQUASHME! [VPG]: drm/i915: Add fences to scheduler dump
When things go wrong with native syncs, it can be useful to see what
the scheduler/sync connection is. This patch adds the native sync
status of each node to the scheduler's status dump.

This patch could be squashed in to the one that originally added the
scheduler status dump facility:
  commit 557478d5fbdded9a79fd09c6f021ffbdb9dfe9ce
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Thu Apr 24 15:40:39 2014 +0100
  Subject: FOR_UPSTREAM [VPG]: drm/i915: Added debug state dump facilities to scheduler
  Change-Id: I0634168e3f3465ff023f5a673165c90b07e535b6

Change-Id: I5c9928c47b9cb68bce7788e60f2ef7fb6bf27883
For: VIZ-1587
For: VIZ-4741
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>