SQUASHME! [VPG]: drm/i915: Proper per ring timelines for legacy mode
The native sync implementation was updated recently to work with the
GPU scheduler. That change was a little confused and ambiguous in
legacy mode. This patch makes legacy mode match execlist mode of
having a timeline per ring per context. This removes any ambiguity of
which ring a given timeline is in use on. It is also necessary for
future work which requires matching timelines to rings when they are
in use.

This patch can be squashed into the patch which made the per context
change initially:
  commit 7b04cfcfcb08fa5f33d61a0870cf0337f5cc2259
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Fri Jan 9 13:32:36 2015 +0000
  Subject: REVERTME [VPG]: drm/i915: Make native sync work with scheduler
  Change-Id: Ia80f332a25c2a3dfcaaadcf4c8bfcf718f24cf64

Change-Id: I3406bd3f9c62dd2fa81208a8f2ff1ace2be1e622
For: VIZ-1587
For: VIZ-4741
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>